[IMP] web_gantt: improvements of drag and drop feature
The main improvement brought by this commit is the
possibility to copy a pill by dragging and dropping it
with the ctrl or the command key pressed.

Along the way we have reviewed the grid and the
containment options used by jQuery UI.

Now it is no more possible to drag a pill
beyond the gantt left boundary (i.e. over the row
headers). Sadly, due to a jQuery UI bug, there is no way
to do the same for the resizing of pills.

The reactivity to the window resizing has been improved.

The style has been rewiewed also a lot to offer a better
readabilty and reactivity of the gantt view.

Task ID: 2067422

closes odoo/enterprise#5656

Signed-off-by: VincentSchippefilt <VincentSchippefilt@users.noreply.github.com>
Co-authored-by: Stefano Rigano <sri@odoo.com>