[FIX] hr_appraisal: Avoid infinite employee creation loop
If you have an empty database, without any employee:
Try to create your first employee.

But by default "Appraisal by managers" is checked,
hence the appraisal managers field is required.
You try to set a manager, but there are no employee!
You then want to create an employee to be the manager.

But by default "Appraisal by managers" is checked,
hence the appraisal managers field is required.
You try to set a manager, but there are no employee!
You then want to create an employee to be the manager.

But by default "Appraisal by managers" is checked,
hence the appraisal managers field is required.
You try to set a manager, but there are no employee!
You then want to create an employee to be the manager.

But by default "Appraisal by managers" is checked,
hence the appraisal managers field is required.
You try to set a manager, but there are no employee!
You then want to create an employee to be the manager.

   ---->
  ^     |
  |     |
  |     v
   <----

To solve this infinite employee creation loop,
the appraisal managers field should not be checked by default.
For consistency, the appraisal by employee should not be checked.