[IMP] industry_fsm: remove default user_id if in fsm app
In fsm, we have a common use case which is someone planning tasks for other (someone is responsible of dispatching, others go on the field). So that person is not the one who will perform the task and therefor not the one that must be assigned to.

However in some views this make sense (like in My Tasks), so we override the project behavior only or some action using a default_ context key.

task-2047389         closes-5388