[FIX] delivery_usps: error message when quantity exceeds max
USPS delivery only allows 999 as max value for the Quantity field [1]. The
default error message by the api is not very intuitive so we catch it
instead and check if the picking has lines that have quantity greater
than 0. We give the user more intuitive message than the one provided by
the api.

Additionally, we also consider the minimum value for weight fields
(which is 1 ounce).

Source:
[1] https://www.usps.com/business/web-tools-apis/evs-international-label-api.htm

closes odoo/enterprise#5452

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>