[FIX] pos_loyalty: selection fields not working
Since the ORM refactoring, it is not possible anymore to declare a
Selection field with a tuple containing the selection values, but
we have to use a list.

So we modified the declaration of fields 'rule_type' and
'reward_type' to match the correct syntax for Selection fields.

also changed the layout of the reward_rule to make the fields rule_type and reward_type
readable.
Variable label for point_cost field to give a better understanding of
it's function.

closes odoo/enterprise#5525

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>