[IMP] web_map: add list of records on map views
Before this imp, when activating the map view, the user had no idea which
records where there, but not displayed because either they have no address
or their address cannot resolve into coordinates.

After this imp, the user has a clear idea about the records that are
available, displayed and not

This also adds the paging capability to the map view.

closes odoo/enterprise#5541

Task-id: 2066736
Signed-off-by: VincentSchippefilt <VincentSchippefilt@users.noreply.github.com>