[IMP] planning, hr_holidays: gantt unavailability
Adapt the gantt unavailability methods from planning and hr_holidays
to the move of related methods ( _get_work_interval() and
_get_unavailable_intervals() ) from hr.employee to resource.resource in
community. This move was needed to allow the unavailability to be
displayed for users without related employees (like in FSM).

Also fixes those methods, which were not displaying correctly the
unavailabilities when multiple group bys where used.

Task-2069861