[FIX] industry_fsm: replace fsm_state by fsm_done
Since the Field Service flow has been orthogonalized and linked more
closely to the Project one, the fsm_state variable became quite
redundant and potentially not in adequation with the project flow.

So we remove this selection field and replace it with a simple
boolean, to simplify the whole flow.

task-2047389         closes-5388

closes odoo/enterprise#5388

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>