[FIX] delivery_*: warning for unresponsive API
The external delivery tests crash on a daily basis due to timeout
errors (bpost) or when the delivery address match a temporary
unavailable country (usps).

We decided to abort the tests prepromptly for such errors as the tests
are value-less in such cases. The original error is still logged as a
warning.

closes odoo/enterprise#5547

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>