[FIX] documents: upload documents via mail alias
PURPOSE:

Sending emails to the document mail alias used to crash, violating
unicity constraint for attachment_id on documents. This was caused
by the creation of a document for the incoming email, to which the
attachment was linked, a proper document was then generated with
the same attachment_id, violating the constraint.

By overwriting message_post, we can pass it no_document=True in
the context such that the attachment_id is not set on the email
document, thus preventing the conflict later in the flow.

part of task-2067097

closes odoo/enterprise#5555

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>