[FIX] sale_ebay: add a cron to synchronise stock in order-based synchronisation
Commit d36a1bf76bd introduced a new synchronisation mechanism based on orders.
To be more efficient (as the old system could run into the API limits),
it only updates products based on orders sent by eBay.

As a result, if some product that synchronises its stock with eBay gets
replenished or sold within Odoo, eBay is not updated.
A separate cron needs to run to fill that need; this cron can run independently
and as often as needed by the user.

Note that it will perform N calls each time, where N is the number of products
for which the eBay quantity has changed from the virtual_available quantity
since last run.

opw 2070413

closes odoo/enterprise#5592

X-original-commit: 19eefa73c5432de6e22b3ac6d562fcf65a5f8855
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>