[IMP] hr_referral: Improve characters names
Purpose
=======

On the referral app, four entities or "creatures" are presented:
- Mystic
- Aboriginal
- Cyborg
- Dog

Now, the fifth one "girl" is presented as an equivalent for those 4, and that can be seen as problematic.
The problem is not having a girl as an avatar to chose from, but the equivalence that is implied.

An example is: Imagine that you have four avatars to chose: stupid, idiot, dumb, and Odoo consultant.
An odoo consultant might feel offended by that - even if it is a joke.

Since this is an app that we are rolling out for millions of users, I really believe that we
should be on the safe side of not offending anyone.

On top of that - Aboriginal are real people. They are a minority that the app sets as an equivalence
to a cyborg, mystic, and dog. That can also be seen as offensive.

If that gets on the internet in a wrong way can be really bad publicity for Odoo.

closes odoo/enterprise#5579

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>