[IMP] account: manual reconciliation
Improve the manual reconciliation widget:
When coming from the aged receivable or the partner followup and clicking "reconcile" we want to be able to see all entries for that partner that are not reconciled (before we only showed something if they were at least one debit and one credit entries for that partner). We added a new context key (all_entries) for that use case.

closes odoo/enterprise#5589

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>