[FIX] account_invoice_extract: set other fields when vendor reference is duplicate
PURPOSE:

Vendor references on vendor bills are supposed to be unique, previously
this would prevent the OCR from filling any field if an existing vendor
bill had the same reference as the one currently being processed. This
fixes that issue by adding an error status for duplicated vendor
reference.

When this occurs, we set the extract_status_code, clear the vendor
reference from the record and store it in duplicated_vendor_ref so that
it can be shown to the user in a warning banner.

NOTE: It has been decided that codes from 100-199 would be reserved for warnings.

task-2067097