[REF] web_enterprise: breadcrumb and cp_searchview flexible width
This commit is linked to community changes:
"1) The size of the left part of the control panel will take free
space left. Normally, the width is always fixed to 50/50 but
now the breadcrumb can take more or less space if needed.
It will be useful on mobile where the search bar is dynamically
sized."

Because of this we can now remove "full" and "not-full" css rules.
On mobile the search bar is transformed into a fa-search button
which takes less space. So breadcrumb will be larger by taking all
the available width.

Task ID : 1924760