[IMP] web: SearchView: Allow user to quickly search on mobile
Currently the search panel is nice for a search with filters but
that was not 100% efficient for simple research without filters.

The Search is now integrated in the form view, with the ability
to open the Search panel to apply filters or group by.

Most of the work has been done in community and we simply have to
adjust the design in here.
We also had to remove some tests as it's already tested in ent.

Task ID : 1924760