[FIX] web_studio: show only value and not label for inline field
In report editor, when an inline field is dropped, it sometimes
shows label for the field depending on where it has been dropped
(for example in empty `col` inside a `row`), and only value in
other cases.

This commit makes the behavior consistent by always dislplaying
only value of the field while dropping an inline field.

task-2057447

closes odoo/enterprise#5811

X-original-commit: a94bf5226edb83f201b2d23a646c96194afe72fc
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>