[FIX] web_enterprise: mobile buttons dropdown for custom FormRenderer
Custom FormRenderer overwrites the MobileFormRenderer set on mobile
devices. This deactivates the grouping of the control panel's buttons
into a dropdown (e.g. in InvoiceExtractFormRenderer).

This is due to the fact that, in mobile, we use an `include` on the
FormView to set the MobileFormRenderer as Renderer but the original
FormRenderer is not modified. Then any custom FormRenderer overwrites
the MobileFormRenderer, removing the code used to fold the buttons into
a dropdown.

This commit modifies the way the MobileFormRenderer's behavior is set by
directly using an `include` (on mobile only) on the FormRenderer. It
then makes the FormView's changes useless.

Task ID: 2056976

closes odoo/enterprise#5636

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>