[FIX] l10n_mx_reports: Report filename must be uppercase
Some Mexican reports are required by SAT to be uploaded to their
online platform. These XML reports need to have the filename in
uppercase characters.

Before this commit users needed to download the XML report files to
thier local machines, and then renamed it to use capital letters,
because when they uploaded the file with the name Odoo generated
(lowercase) then they got an error from the SAT platform rejecting
the filename.

Here we are making the reports to be generated automatically in
uppercase in order to avoid users to do additional steps.

Closes #5685

closes odoo/enterprise#5716

X-original-commit: fb1a82fb41629dd6ddf57f20ce5af0a5307fce7f
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>