[FIX] helpdesk_timesheet: fix access right issue
Before this commit, if helpdesk_timesheet is installed and user has
only project user access rights, opening the Project App creates an
access right error on helpdesk.ticket, because of the call to
ticket_count in Project kanban and form views.

This commit fixes the issue by setting the Helpdesk user group on the
problematic view inheritances.

Task-2066533