[FIX] helpdesk: multiple installs inside a single request
The modification of a helpdesk team may trigger the installation of some
modules.  The code was calling multiple times the installation of
modules (once per module to install), and this was triggering cache
misses on computed fields (probably because of a completely inconsistent
cache/registry).  We have grouped all the modules to install into a
single installation, and this fixes the issue.  It also makes the code
more generic.

Co-authored with Samuel Degueldre <sad@odoo.com>

Task 2070454

closes odoo/enterprise#5648

Signed-off-by: Raphael Collet (rco) <rco@openerp.com>