[FIX] web_grid: avoid reload after cancel a modal windows
— In the 'Timesheets' module;
— In 'Timesheets to validate';
— Click on 'Validate';
— A modal windows opens;
— Close the window by clicking on 'Cancel'.

Before this commit, an error is arisen, this occurs because when closing
the modal, the code will try to update the grid, but without the correct
parameters which arise an error.

Now, if we close the modal by clicking on 'Cancel', the update of the
grid is avoided.

opw-2065474

Fine-tunning of b4f028df3681f776e9697d1311ec091c279f8a94 :
$.when is deprecated

closes odoo/enterprise#5650

X-original-commit: cf6bde4
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>