[FIX] l10n_mx_edi_external_trade: Define ns in the Comprobante
The SAT documentation indicate that the schema for the complements must
be in the node "Comprobante", and now the PAC makes this validation. The
do not allow stamp the CFDI if the schema is defined only in the
complement.

Now is added in the Comprobante

But, to get this is necessary the next:

Allow add elements with namespace in the CFDI template with xpath

In the next case:
Is necessary define the next attribute in the CFDI in the Comprobante
level:
xmlns:cce11="http://www.sat.gob.mx/ComercioExterior11"

If you try with:
<attribute
name="xmlns:cce11">http://www.sat.gob.mx/ComercioExterior11</attribute>

The next error is returned:
Invalid attribute name 'xmlns:cce11'

For this reason, will be defined with:
<attribute name="xmlns__cce11">http://www.sat.gob.mx/ComercioExterior11</attribute>

And the __ will be replaced by : when the CFDI is generated.

closes odoo/enterprise#5676

X-original-commit: c648af3c4c92922560bc52b127fd6c70280d158b
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>