[IMP] *: remove image_64
Enterprise counterpart of https://github.com/odoo/odoo/pull/36147

There are too many image sizes. Since they are stored resized this takes time to
generate when saving a new image, it's more rows on the attachment table, more
files on the disk, ...

64px is close enough to 128px that it can be removed without a big impact on
download size.

It will even reduce download and number of requests when both images are
displayed because now only one has to be downloaded and then benefit from cache.

The difference between the two is typically around 1.5kB which is negligible
these days, especially when the request overhead is around 0.5kB already, not
even taking into account other factors such as latency.

If a 64px image must absolutely be returned, it is still possible to pass the
size parameters to the image route. But the current guideline is to handle
resizing in the views when necessary.

Views
=====

- remove width and height attributes when existing CSS rules are overriding them
  (eg. `.oe_kanban_avatar` in the right context)
- add CSS rules instead of width and height attributes when possible
- use `object-fit: cover;` when width and height are forced to avoid distortion
  of non-square images
- for products, use `object-fit: contain;` instead, same reason but without crop
- add new CSS rules when the expected size was max 64px*64px before due to the
  image size itself

Note: forcing width, height and fit does not work in combination with img-fluid.

task-2060865

closes odoo/enterprise#5296

Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>