[IMP] planning: made some improvements and fixes on the recurrence feature
-IMP: recurrence can now be modified

Rather than just simply deleting the recurrence when changing
the value of some of its field (eg. change a repeat forever to
a repeat until <date>) we now interrupt it to the record that is
been changed a create a new one from it.

Diagram: A - A - A (user stops recurrence A here)

                 |

		 B - B - B (recurrence B)

NB: planning slots that are already published are not removed
from any recurrence wheter its deleted or modified.

-FIX: moved the recurrence logic in the inverse method of the repeat fields.
-FIX: looped on self when changing recurrence parameters as self can be multiple records.
-FIX: when we change a recurrence parameters, we change last recurrence parameters
      to 'until' and sets its stop date to the currently edited end date so it's
      not re-generated when the cron runs.

Task-2067854