[FIX] industry_fsm: button visiblity, probably not the last one
This commit reviews some attrs of FSM button in FSM app and project
app. Indeed, accodring to the checked options on the project, some
criterions should (or not) be taken into account to display an action
button. Moreover, some button needs to be displayed in primary or
secondary style, which can make the attrs complex.
See related task for the desired behavior.

Task-2061910