[FIX] account_report: adapt test to work around code logic
The test 'Account report m2m filters' randomly fails between the two
dropdown clicks.  Apparently, the account report code is really weird.
It reloads everything between clicks, then rerender himself completely,
and rerender the search view as well, then tries to reopen stuff that
should be open.

So, this relies on the fact that some other code (in this case,
bootstrap) reopen the dropdown. But there is a small period of time in
which the dropdowns are closed.  It is extremely likely that this is the
cause of the random crash: most of the time bootstrap is first and
reopen the dropdown before the test, but sometimes, this is the
opposite, and in this case, the test fails.

This fix is not really a fix. We simply wait a little bit more to make
sure the dropdown is open whenever the test clicks on it.

However, a real fix is to rewrite the whole thing.  It is an bunch of
imperative complicated code (and with server side rendering as well).

closes odoo/enterprise#5689

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>