[FIX] quality: get team from the right company when creating checks
Use quality in multicompany.
The default team for checks is searched without check on the company.
As a result, workorders can be created with a team_id being in another company.
It follows that MOs with such workorders cannot be planned or validated.
Instead, we raise to avoid the user getting stuck and needing support's help.

opw 2072587

closes odoo/enterprise#5744

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>