[IMP] industry_fsm_report: worksheet duplication
Since a worksheet generates its own model, when duplicating one of
the them, you will get an SQL error (since we can not have 2 ir.model
with the same name).
This commit changes the code in order to allow duplication with a
different name. The newly generated model will have the id of the
worksheet template duplicated.

Task-2074002