[FIX] delivery_*: remove traceback from warning
The runbot is an automated building agent and a monitoring tool, it is
used by a github bot "robodoo" to build and test every pull request to
ensure any modification in the code base doesn't break any component of
the application.

Every build is colored according to its logs:
* green: no warning, no error.
* yellow: warning but no error.
* error: error (traceback or failed test)

There are some tests that are known to be non-deterministic, false
positive or obsoletes. Those tests are explicitly flagged in the source
code and are not ran as part of regular build. Those tests are only ran
once everyday during the night. As part as those tests are the
"external" tests, they are integration tests with remote APIs and
sometime fail due to error API-side. e.g. Belgium's postal service
"bpost" API is unresponsive on a regular basis.

As the tests fail, the runbot is colored red. As it is red all the time,
the team in charge of monitoring builds did not oversee the logs
anymore and some real errors were introduced.

To ensure non-deterministic tests do not color the runbot red, timeout
and related errors have been catch and the test aborted with a warning
and related traceback. The problem is that the "Traceback" string is
matched by the runbot and the build colored red which fails the intended
purpose.

As the catch is very precise (specific exception + regexp on payload),
the related traceback is no more logged as part of the warning.

closes odoo/enterprise#5713

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>