[REM] account, project, sale, *: remove dead code 'return_label'
Since Odoo 9.0, the `return_label` in url was not working anymore.
What it did before 9.0 is add an item in the breadcrumb which once clicked
would redirect to the frontend.
This was handled in the `do_load_state` of the action manager in JS.

We don't need to reimplement that behavior as now we have the 'Go to website'
stat button. Thus we remove the code.

Coming with https://github.com/odoo/odoo/pull/36325

closes odoo/enterprise#5725

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>