[FIX] sale_subscription: fix failure to create subscription alert that updates fields
Before this commit, subscription alerts that updated field values could not be created due to lack of access rights. For instance, a subscription alert with "Mark as To Renew" as action would raise an error on creation.

This commit makes field actions being created in sudo as we need to write on ir.actions.server.
It also fixes the domain and pre-domain of the action failing to match any record.

task-1998636