[FIX] project_enterprise: avoid crash on unassigned creation
When grouping the gantt view, we want to always see the "unassigned tasks"
(where the user_id field is not set), even if there is no record. The js_class
of gantt view does the job, but when adding record on those "fake groups", it
crashes as we don't have the default value of the groupby fields.
This commit provides those missing values by determining them from the "path".

Task-2074002