[IMP] web_map: handle records of same co-ordinates in marker
Before this commit,
Records containing the same co-ordinates will create markers which are
overlapped with each other.

After this commit,
Records containing the same co-ordinates are displayed in one marker with badge
having a count of duplicate co-ordinates records. Clicking on 'Open' button in
popover of this marker will open a list view which contains records of same
co-ordinates.

task-2009017