[IMP] social: add field on social media to ease accounts linking
´´can_link_accounts´´ now defines whether an account has to be linked manually.
It is used for push notification where customers do not create account. They
use a default account coming with website.

´´has_streams´´ previously was used to both account creation and stream
creation. It was used only for push notifications. With linkedIn we need
to be able to separate account linking from stream creation. That is why
we introduce the new field allowing to separate both concepts.

Task ID 2066644