[FIX] mrp_mps: inventory quantity in starting inventory row
Usecase to reproduce:
- Create a new product -> do an inventory adjustement
- Open MPS -> The quantity added is in row Actual Demand and it should
be in starting inventory.

Same process for decreasing quantity in Actual Replenishment.

It's due to commit da33eb36b80d1b2a0004c28cf20d6b0263bcb643 it's
possible to directly edit quant that will create a move from inventory
but without a stock.inveotry record. In this case it should be
in the starting inventory and not Actual Demand/Replenishment.