[FIX] mrp_workorder: store uom on quality check
Since fdc03302538b6541e0b4379c73dac83c9cb62671, the Unit of measure
on a quality check is related from the workorder line and not from
the product anymore. The issue is that the workorder line are deleted
at the end of the production. The UoM is lost when the production is finished

This fix uses the UoM of the move (it's the same than the bom line one so
the commit fdc03302538b6541e0b4379c73dac83c9cb62671, is still valid) instead of the
one on the workorder line

closes odoo/enterprise#5741

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>