[FIX] industry_fsm_report: allow module uninstallation
When uninstalling a module, the ORM check all the xml ids, and remove them
by inverse order of creation date. First the demo, then the data, the views,
... and finally models and fields.
Now that fsm report has fields defined in its demo data, a CacheMiss exception
raises as the fields are implicitely removed; they are in database but not in
the ORM cache. This came from the fact that during the uninstallation, we
should let the ORM completely handle the removal of module data (see `unlink`
of project.worksheet.template).
To be sure everything is well removed, this commit also creates some xml ids.
We now have
- the manual deletion of a worksheet: the removal of its linked actions, views,
accesses, ... are explicitely handled in 'unlink'
- the module uninstallation: handle by the ORM

Purpose is to allow the module uninstallation, but also be sure all generated
data are removed (no database pollution).

Thanks to Elkasitu for his investigation !

Task-2074002