[IMP] industry_fsm_report: demo data generation
A new worksheet template has been added to Field Service Worksheet
demo data. But if you try to print or preview this report, you won't
get the fields contained in it.

This commit fixes the issue by adding a name and type to the view
created in demo data.

Task-2074002

closes odoo/enterprise#5743

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>