[FIX] sale_subscription: wrong action created when adding an sms/mail alert
The 'base.automation' record that was created was wrong,
its action was set to "Update the record" and the SMS/mail was never sent

This is because activity user_ids was never False and therefore
email and sms actions could never be created because the action
was considered an activity each time.
So i inverted the last 2 elif in _configure_alert_from_action.

closes odoo/enterprise#5633

Task-id: 2070836
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>