[FIX] sale_ebay: remove synchonisation at unlink
Backport of 0349d66018b6319a5200394d1ff36cfc980f04f9

If the customer is using the new synchonisation introduced by d36a1bf76bd1cd8
this would trigger an explicit call to the old synchonisation method which would
create duplicate transactions for orders with more than one product.

Note that this call was introduced to avoid missing the synchonisation of one
transation that would have occurred between the last run of the cron of the old
synchonisation and the moment the unlink button is clicked; this would be
between 5 and 10 minutes.
But the old synchonisation is flaky, so it missed synchronising products anyway;
furthermore it only called the cron with sync_big_stocks=False, so it would also
not work if there were more than 150 items listed.
So removing that call leaves a gap, but which was already existing.