[FIX] marketing_automation: fix some usability issues
In the marketing automation app, when we copy a campaign, do not copy its
activities "require syn" flac. Indeed copy of compaign is in draft state.
Fix by adding explicitly copy=False on the require sync field.

In order to be generic with sms and mailing and coherent with current
naming, rename the string of mass_mailing_id field from "email template"
into "marketing template"

Task 2075733

closes odoo/enterprise#5765

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>