[FIX] documents_account: fix main_attachment_id not set on vendor bill creation
When creating a vendor bills from the Documents app, the
'message_main_attachment_id' field wasn't set on the record. This
happened because the message posted in the chatter was linked to the
wrong 'res_model' which prevented it from being selected as the main
attachment.

This was causing a bug in the OCR module ('account.invoice.extract')
where we use this field to determine which attachment must be sent to
the OCR server.

closes odoo/enterprise#5768

Signed-off-by: Florian Daloze (fda) <fda@odoo.com>