[FIX] stock_barcode: async load on nomenclature
is_product use parse_barcode that use the nomenclature rule loaded
in load(). The abstract client action for barcode call BarcodeParser
load in willstart but do not wait for the resolve then if a product
is scanned before then it will raise due to .lenght on nomenclature
rules since they are not loaded.

closes odoo/enterprise#5772

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>