[IMP] sale_ebay: better support for all eBay sites
To be able to easily change the default eBay website to use,
the mapping from code to all available sites to date is added to the data.
Before, it was not even clear that other sites are supported by the module;
moreover it required a server action to configure everything.
Since other sites are added, we modify the default to make sure it stays
on the main site (since country, currency and site should stay coherent).