[IMP/REM] sale_ebay: make order-based synchronisation the default, remove product-based synchronisation
The order-based synchronisation is the right way to use the eBay API.
It only queries the API for the recent orders, and updates products accordingly.
The old synchronisation was making queries for each product,
which was explicitly discouraged by the API, and went against rate limits for
stores with hundreds of products.
Start from commit d36a1bf76bd for more details.

closes odoo/enterprise#6912

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>