[FIX] sale_ebay: add offline test to check processing of orders
This adds a test that the result from a request to eBay is normally processed.
This could reveal if e.g. a change in ORM modify the expected format of dates.
It can also serve as an entry point to understand the working of the module.
The test data follows the exact same format of a real reply, anonymised.