[FIX] sale_subscription: wrong discount in upsell
Enable pricelists (Prices computed from formulas) and discounts.
Create a pricelist for a product [DEMO] with a fixed price showing the discount
to the customer.
Now create a subscription and edit the next invoice date to a future
date (example, one more year). Put the state to "Upsell". Create an
Upsell including [DEMO], specifying a [DATE] which is less than the next invoice
date minus the renewal period but still greater than today (just for the
sake of the realistic use case) and click on "create and view quotation".

The discount will be wrong. This happens because the amount to discount
is calculated on the basis of the remaning period up to the next invoice
date (to have a 'fair' invoice).

To get the elapsed period the last invoice date is computed as:

last_invoice_date = next_invoice_date - invoicing_period

and the discount will be calculated as

discount = ([DATE] - last_invoice_date)/(next_invoice_date-last_invoice_date)

This is based on the assumption that
last_invoice_date < [DATE] < next_invoice_date

so for example when the next_invoice_date is pushed forward in the future
without touching the invoicing_period the last_invoice_date is not correctly
calculated and the discount will be wrong as well, because
[DATE] < last_invoice_date < next_invoice_date

A solution is to look if the ratio is out of the range [0-1] and
if it is returning a standard ratio as well as an informative message to
be displayed on the sale order to inform the user of what is happening.

opw-2074068

closes odoo/enterprise#5969

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>