[FIX] planning: fix creation of shift template without role_id
When creating a template in planning, a check is made that this template doesn't already exist (the same combination of role, and time).
When extracting the role from the result of the read group, it was expected that the value would have the form [id, name], but when the relation is not set it is a False boolean.
This commit fixes the issue by checking the value is not falsy before extracting the id.

Task-2036357

closes odoo/enterprise#5849

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>