[FIX] planning: correct values in reporting
In planning, we want to see the number hours per user on a
certain period (month, week, ...). For planning shift on several
days, we should simply divide the allocated hours by the days span
of the shift, without taking the leaves and working calendar into
account.
'working_days_count' is only used in project_timesheet_forecast to
compare timesheet and forecast with leaves into account.

Task-2077822