[FIX] web_gantt: reflect access rights
When the current user has not access to create a record, the "add"
button should not be displayed. This is due to the fact that the
'create', 'edit' and 'delete' attributes are not set on gantt tag in
the arch.
This commit sets those parameters according to the access of the user,
like the other view.