[FIX] web_enterprise: hide arrow in burger menu from the app switcher
Before this commit, the user was thinking that he can click on his
username when the burger menu was opened from the app switcher.
It was due to the fact that an arrow was present.
Note that this dropdown is usefull when you need to switch between
burger menu contents: app menu or user menu.

In this case, user menu is the only one that can be shown.
To fix this, we simply hide this arrow when the burger menu is
displayed from the app switcher.

opw-2058510

closes odoo/enterprise#5859

X-original-commit: 3bd392aafe7ae9c151b2a72ffbb74c6e6f851881
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>