[FIX] marketing_automation(_sms) : view bugfixes
- In the marketing automation app,
the Marketing mailing form had a required and invisible
field inherited from a parent view.
Remove it only for the marketing automation.

- In the form of a marketing activity, create a mailing (sms/mail)
from the Many2One field and save the record, the field is not
correctly updated. Fix this issue by removing name invible field of the
form (because the name was taken firstly by the default widget instead
of the display name which is subject now).

- In the marketing automation sms,
the sms marketing form had a field for sms template.
Make it invisible to be consistent with the sms app.

- When we create a mass sms from marketing automation,
it was accessible also from the sms marketing app.
Force the domain (in sms marketing) to fix this issue.

TASK_ID : 2075733