[FIX] account_invoice_extract: fix partner selection from name
When trying to find a partner in database matching the name of the
partner detected by the OCR, the name found by the OCR is split, no
restriction are applied on the length of those splits. This could lead
to splits such as a single letter which would match many partners in the
database.

closes odoo/enterprise#5841

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>