[IMP] hr_payroll: Remove regular pay structure
Purpose
======
The boolean field `regular_pay` on payslip structures is mainly used to retrieve
the default structure of a structure type. This is a problem because a struture
type could have multiple structures with the checkbox set to True.
How to decide which is the default one then ?
Currently, it takes the first found.

There should be only one default structure for any structure type.

Specification
==========

Remove the `regular_pay` field on the payslip structure.
Make the default structure (field `default_struct_id`) of a structure type a
regular many2one (currently computed).

Adapt data and tests.

closes odoo/enterprise#6785

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>