[FIX] industry_fsm_stock: display correct quantity on pickings.
When creating a sale_order with a service product configured to create a task in an existing project, this project can be an fsm project, the task can then follow the fsm flow.
What's special with starting the flow this way is that the sale order linked to the task is already confirmed (needed to generate the task).
Therefore when marking the task as done, the hook function _action_launch_stock_rule doesn't run again.
However it is called when we add a sale_line to the sale order, but not when we write on that line (_action_launch_stock_rule isn't called if the line is already sold which is the case here).
This lead the pickings of this task to have it's delivered quantities stuck at one. To fix this we manually call _action_launch_stock_rule when marking an fsm task as done.

There was also a usability problem due to validating all pickings upon mark as done, for products with tracking, the user has no way to set bar codes/lots etc... Those product's picking would have been false.

To avoid that situation, we hide those kind of products from the fsm flow.

Task-2077576