[FIX] planning: 'False' in mail template
When sending a mail template with the light layout set, the layout will
create a cache record of a mail.message and display its 'record_name' in
the header. Here we are sending a 'planning.planning' email which as a 'name'
field never set, resulting as 'False' as string value.
Since the field is not used, we prefer removing it, but to avoid having
'planning.planning,4' instead in the mail layout header, overriding the
magic  `_compute_display_name` is required.

Task-2077822

closes odoo/enterprise#5848

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>