[FIX] helpdesk_account: fix refund modal
Before this commit, and since compta-pocalyspe, the wizard opened by clicking on
"Refund" button on tickets was causing an error, or was hidden and
pre-filled with a move chosen by an accounting business logic related
way. This wizard is inherited from account module and didn't enjoy all
the changes made there.

This commit fixes the issue by ensuring that no move is selected in
the default_get method of the wizard if it is opened in a helpdesk context.

Task-2047999

closes odoo/enterprise#5857

Signed-off-by: gla-odoo <gla-odoo@users.noreply.github.com>