[FIX] sale_subscription: studio crash on new view
Install subscription (sale_subscription). Create a new subscription with
any template, put the state in "Upsell", create an upsell and generate a
single quotation ("Create and View Quotation"). Now From the quotation,
click on "Subscription", there should be only one. Now open studio and
activate list view.

Javascript traceback will popup. Moreover the view of the subscription
module will be altered in a faulty way (can be recovered from "Window
Actions", looking for the correct action and removing the last added
tree from "View Mode").

This is caused by studio not detecting all the active views, thus giving
the possibility to add an already active view, because when there is
just one element to display all other views are filtered out from the
python code and the frontend see only the current view.

Changing the way the form view is retrieved in case of a single document
fix the issue.

opw-2076241

closes odoo/enterprise#5958

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>