[FIX] web_enterprise: cannot open discuss channel on iOS 13
Since iOS 13 when the user clicks to open a channel (from the discuss'
systray menu, for example), nothing happens.

In fact, the "touch" event (used for the swipe actions) is triggered but
not the "click" event (used to open the channel).

This issue is only reproducible on Safari on iOS 13 (currently 13.1.2).

This commit works around this issue by replacing the swipe action
`button` element by a `div` (which doesn't seem to trigger the issue).

The styling had to be adapted in consequence.

closes odoo/enterprise#5885

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>